pyproject.toml
[project]
name = "fastpath"
version = "0.1.0"
description = "Extensible pathology whole-slide image viewer with preprocessing-first architecture"
requires-python = ">=3.11"
dependencies = [
    "PySide6>=6.6",
    "openslide-python>=1.3",
    "numpy>=1.26",
    "Pillow>=10.0",
    "scipy>=1.11",
    "rtree>=1.0",
    "click>=8.0",
    "tqdm>=4.0",
    "pyvips>=3.1.1",
    # Note: pyvips-binary removed - using system libvips with OpenSlide support
    # Install libvips: https://github.com/libvips/build-win64-mxe/releases (all variant)
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-qt>=4.0",
]
vips = [
    "pyvips>=2.2",
]
rust = [
    # Build with: cd src/fastpath_core && maturin develop --release
    # Note: fastpath_core is built separately with maturin
]
all = [
    "pyvips>=2.2",
    "pytest>=8.0",
    "pytest-qt>=4.0",
]

[project.scripts]
fastpath = "fastpath.__main__:main"
fastpath-preprocess = "fastpath.preprocess.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fastpath"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "maturin>=1.11.5",
    "pytest>=9.0.2",
    "pytest-qt>=4.5.0",
]
